refactor(cleanup): split cleanup.go god-object into strategy files (#227)
cleanup.go was 948 LOC mixing 8 cleanup strategies + DB maintenance. Split
along strategy boundaries:

- cleanup.go (377 LOC, was 948): Manager struct, constructor, config
  setters (Set*), Run loop, RunOnce dispatcher, activeCameraIDs, and the
  shared BatchDeleteRecordingsWithFiles + adaptiveBatchSleep core helpers
- time_retention.go (196 LOC): timeBasedCleanup + archived/health/
  transcode-history retention
- disk_threshold.go (67 LOC): diskThresholdCleanup
- dark_segments.go (43 LOC): darkSegmentCleanup
- orphans.go (89 LOC): orphanFileCleanup + cleanOrphansForCamera
- repair.go (147 LOC): staleRecordCleanup + fixStaleMJPEGRecords +
  repairZeroDurationRecordings + probeDuration
- db_maintenance.go (125 LOC): performDatabaseMaintenance (WAL checkpoint
  PASSIVE→TRUNCATE escalation + incremental vacuum) + updateSQLiteMetrics

No behavioral change: all methods moved verbatim with original comments,
all signatures preserved. The RunOnce dispatcher fans out to the per-
strategy files unchanged.

Verified: go build passes; go test -race ./internal/cleanup/... passes;
coverage 51.3% maintained (split is mechanical; new edge-case tests to
reach the 65% target are a separate follow-up per the issue); go vet +
gofmt clean.

Closes #227
